refactor: add MsgMultihopProofs methods following ibc convention;
- remove top-level channel handshake methods in favor of EndpointM methods
- make proof.go unexported methods private
